requirements.txt
# This file is part of WellPlanned AI.
#
# WellPlanned AI is free software: you can redistribute it and/or modify
# it under the terms of the GNU Affero General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# WellPlanned AI is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
# GNU Affero General Public License for more details.
#
# You should have received a copy of the GNU Affero General Public License
# along with WellPlanned AI. If not, see <https://www.gnu.org/licenses/>.

fastapi
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
openai
python-dotenv
beautifulsoup4
pypdf
langchain-google-genai
python-multipart
PyMuPDF
pillow
httpx==0.27.2
supabase
selenium
unstructured
